requirements.txt
asgiref==3.8.1
Django==4.2.13
gunicorn==22.0.0
packaging==24.1
sqlparse==0.5.0
typing_extensions==4.12.2
python-dotenv